place all functions to classes
