+++ classes/output/Strategy/StrategySelectorList.php
@@ -10,7 +10,7 @@ use block_tutor\output\Strategy;
 
 require_once("../studentslist_view.php");
 
-class StrategySelectorList implements Strategy
+class StrategySelectorList extends sirius_student implements Strategy
 {
     private $tablist = '';
 
@@ -20,6 +20,22 @@ class StrategySelectorList implements Strategy
 
     public function get_students(): array
     {
-        // TODO: Implement get_students() method.
+        $return_arr = array('students' => array(), 'groups' => array());
+
+        foreach ($groups_arr as $courseid => $val) {
+            foreach ($val as $groupname => $group_data) {
+                $group_students = $this -> getGroupUsersByRole($group_data -> id, $courseid);
+                foreach ($group_students as $userid => $profile) {
+                    $studentname = $profile -> name;
+
+                    $return_arr['students'][$userid]['studentname'] = $studentname;
+                    $return_arr['students'][$userid]['userid'] = $userid;
+                    $return_arr['groups'][$groupname]['name'] = $groupname;
+                    $return_arr['groups'][$groupname]['groupid'] = $group_data -> id;
+                }
+            }
+        }
+
+        return $return_arr;
     }
 }
